requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.4
decorator==4.4.2
Django==4.2.5
django-cors-headers==4.2.0
fuzzywuzzy==0.18.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
install==1.3.5
Levenshtein==0.22.0
moviepy==1.0.3
numpy==1.26.0
Pillow==10.0.1
proglog==0.1.10
pycparser==2.21
python-dotenv==1.0.0
python-Levenshtein==0.22.0
rapidfuzz==3.3.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
